Delete unused loc/ranges iterators
And move the rest of loc/ranges in loclists/rnglists.
