+++ src/rnglists.rs
@@ -2,7 +2,6 @@ use constants;
 use endianity::{EndianBuf, Endianity};
 use fallible_iterator::FallibleIterator;
 use parser::{self, Error, Format, Result};
-use ranges::{Range, DebugRanges};
 use reader::{Reader, ReaderOffset};
 use Section;
 
@@ -10,11 +9,48 @@ use Section;
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct AddressIndex(pub u64);
 
+/// The `DebugRanges` struct represents the DWARF strings
+/// found in the `.debug_ranges` section.
 #[derive(Debug, Clone, Copy)]
-struct RngListsHeader {
-    format: Format,
-    address_size: u8,
-    offset_entry_count: u32,
+pub struct DebugRanges<R: Reader> {
+    pub(crate) debug_ranges_section: R,
+}
+
+impl<'input, Endian> DebugRanges<EndianBuf<'input, Endian>>
+where
+    Endian: Endianity,
+{
+    /// Construct a new `DebugRanges` instance from the data in the `.debug_ranges`
+    /// section.
+    ///
+    /// It is the caller's responsibility to read the `.debug_ranges` section and
+    /// present it as a `&[u8]` slice. That means using some ELF loader on
+    /// Linux, a Mach-O loader on OSX, etc.
+    ///
+    /// ```
+    /// use gimli::{DebugRanges, LittleEndian};
+    ///
+    /// # let buf = [0x00, 0x01, 0x02, 0x03];
+    /// # let read_debug_ranges_section_somehow = || &buf;
+    /// let debug_ranges = DebugRanges::new(read_debug_ranges_section_somehow(), LittleEndian);
+    /// ```
+    pub fn new(debug_ranges_section: &'input [u8], endian: Endian) -> Self {
+        Self::from(EndianBuf::new(debug_ranges_section, endian))
+    }
+}
+
+impl<R: Reader> Section<R> for DebugRanges<R> {
+    fn section_name() -> &'static str {
+        ".debug_ranges"
+    }
+}
+
+impl<R: Reader> From<R> for DebugRanges<R> {
+    fn from(debug_ranges_section: R) -> Self {
+        DebugRanges {
+            debug_ranges_section,
+        }
+    }
 }
 
 /// The `DebugRngLists` struct represents the contents of the
@@ -22,37 +58,6 @@ struct RngListsHeader {
 #[derive(Debug, Clone, Copy)]
 pub struct DebugRngLists<R: Reader> {
     debug_rnglists_section: R,
-}
-
-impl RngListsHeader {
-    /// Return the serialized size of the table header.
-    fn size(&self) -> u8 {
-        // initial_length + version + address_size + segment_selector_size + offset_entry_count
-        self.format.initial_length_size() + 2 + 1 + 1 + 4
-    }
-}
-
-fn parse_header<R: Reader>(input: &mut R) -> Result<RngListsHeader> {
-    let (length, format) = parser::parse_initial_length(input)?;
-    let length = R::Offset::from_u64(length)?;
-    input.truncate(length)?;
-
-    let version = input.read_u16()?;
-    if version != 5 {
-        return Err(Error::UnknownVersion(version as u64));
-    }
-
-    let address_size = input.read_u8()?;
-    let segment_selector_size = input.read_u8()?;
-    if segment_selector_size != 0 {
-        return Err(Error::UnsupportedSegmentSize);
-    }
-    let offset_entry_count = input.read_u32()?;
-    Ok(RngListsHeader {
-        format: format,
-        address_size: address_size,
-        offset_entry_count: offset_entry_count,
-    })
 }
 
 impl<'input, Endian> DebugRngLists<EndianBuf<'input, Endian>>
@@ -91,6 +96,44 @@ impl<R: Reader> From<R> for DebugRngLists<R> {
             debug_rnglists_section,
         }
     }
+}
+
+#[derive(Debug, Clone, Copy)]
+struct RngListsHeader {
+    format: Format,
+    address_size: u8,
+    offset_entry_count: u32,
+}
+
+impl RngListsHeader {
+    /// Return the serialized size of the table header.
+    fn size(&self) -> u8 {
+        // initial_length + version + address_size + segment_selector_size + offset_entry_count
+        self.format.initial_length_size() + 2 + 1 + 1 + 4
+    }
+}
+
+fn parse_header<R: Reader>(input: &mut R) -> Result<RngListsHeader> {
+    let (length, format) = parser::parse_initial_length(input)?;
+    let length = R::Offset::from_u64(length)?;
+    input.truncate(length)?;
+
+    let version = input.read_u16()?;
+    if version != 5 {
+        return Err(Error::UnknownVersion(version as u64));
+    }
+
+    let address_size = input.read_u8()?;
+    let segment_selector_size = input.read_u8()?;
+    if segment_selector_size != 0 {
+        return Err(Error::UnsupportedSegmentSize);
+    }
+    let offset_entry_count = input.read_u32()?;
+    Ok(RngListsHeader {
+        format: format,
+        address_size: address_size,
+        offset_entry_count: offset_entry_count,
+    })
 }
 
 /// An offset into either the `.debug_ranges` section or the `.debug_rnglists` section,
@@ -422,6 +465,60 @@ impl<R: Reader> FallibleIterator for RngListIter<R> {
     }
 }
 
+/// An address range from the `.debug_ranges` section.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct Range {
+    /// The beginning address of the range.
+    pub begin: u64,
+
+    /// The first address past the end of the range.
+    pub end: u64,
+}
+
+impl Range {
+    /// Check if this is a range end entry.
+    ///
+    /// This will only occur for raw ranges.
+    #[inline]
+    pub fn is_end(&self) -> bool {
+        self.begin == 0 && self.end == 0
+    }
+
+    /// Check if this is a base address selection entry.
+    ///
+    /// A base address selection entry changes the base address that subsequent
+    /// range entries are relative to.  This will only occur for raw ranges.
+    #[inline]
+    pub fn is_base_address(&self, address_size: u8) -> bool {
+        self.begin == !0 >> (64 - address_size * 8)
+    }
+
+    /// Add a base address to this range.
+    ///
+    /// This should only be called for raw ranges.
+    #[inline]
+    pub fn add_base_address(&mut self, base_address: u64, address_size: u8) {
+        debug_assert!(!self.is_end());
+        debug_assert!(!self.is_base_address(address_size));
+        let mask = !0 >> (64 - address_size * 8);
+        self.begin = base_address.wrapping_add(self.begin) & mask;
+        self.end = base_address.wrapping_add(self.end) & mask;
+    }
+
+    /// Parse an address range entry from `.debug_ranges` or `.debug_loc`.
+    #[doc(hidden)]
+    #[inline]
+    pub fn parse<R: Reader>(input: &mut R, address_size: u8) -> Result<Range> {
+        let begin = input.read_address(address_size)?;
+        let end = input.read_address(address_size)?;
+        let range = Range {
+            begin: begin,
+            end: end,
+        };
+        Ok(range)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     extern crate test_assembler;
@@ -664,5 +761,250 @@ mod tests {
             .ranges(RangeListsOffset(buf.len()), 5, 0, 0x01000000)
             .unwrap();
         assert_eq!(ranges.next(), Ok(None));
+    }
+
+    #[test]
+    fn test_range() {
+        let range = Range {
+            begin: 0,
+            end: 0xffffffff,
+        };
+        assert!(!range.is_end());
+        assert!(!range.is_base_address(4));
+        assert!(!range.is_base_address(8));
+
+        let range = Range { begin: 0, end: 0 };
+        assert!(range.is_end());
+        assert!(!range.is_base_address(4));
+        assert!(!range.is_base_address(8));
+
+        let range = Range {
+            begin: 0xffffffff,
+            end: 0,
+        };
+        assert!(!range.is_end());
+        assert!(range.is_base_address(4));
+        assert!(!range.is_base_address(8));
+
+        let range = Range {
+            begin: 0xffffffffffffffff,
+            end: 0,
+        };
+        assert!(!range.is_end());
+        assert!(!range.is_base_address(4));
+        assert!(range.is_base_address(8));
+    }
+
+    #[test]
+    fn test_ranges_32() {
+        let start = Label::new();
+        let first = Label::new();
+        let section = Section::with_endian(Endian::Little)
+            // A range before the offset.
+            .mark(&start)
+            .L32(0x10000).L32(0x10100)
+            .mark(&first)
+            // A normal range.
+            .L32(0x10200).L32(0x10300)
+            // A base address selection followed by a normal range.
+            .L32(0xffffffff).L32(0x02000000)
+            .L32(0x10400).L32(0x10500)
+            // An empty range followed by a normal range.
+            .L32(0x10600).L32(0x10600)
+            .L32(0x10800).L32(0x10900)
+            // A range that starts at 0.
+            .L32(0).L32(1)
+            // A range that ends at -1.
+            .L32(0xffffffff).L32(0x00000000)
+            .L32(0).L32(0xffffffff)
+            // A range end.
+            .L32(0).L32(0)
+            // Some extra data.
+            .L32(0);
+
+        let buf = section.get_contents().unwrap();
+        let debug_ranges = DebugRanges::new(&buf, LittleEndian);
+        let debug_rnglists = DebugRngLists::new(&[], LittleEndian);
+        let rnglists = RangeLists::new(debug_ranges, debug_rnglists).unwrap();
+        let offset = RangeListsOffset((&first - &start) as usize);
+        let version = 4;
+        let mut ranges = rnglists.ranges(offset, version, 4, 0x01000000).unwrap();
+
+        // A normal range.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x01010200,
+                end: 0x01010300,
+            }))
+        );
+
+        // A base address selection followed by a normal range.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x02010400,
+                end: 0x02010500,
+            }))
+        );
+
+        // An empty range followed by a normal range.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x02010800,
+                end: 0x02010900,
+            }))
+        );
+
+        // A range that starts at 0.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x02000000,
+                end: 0x02000001,
+            }))
+        );
+
+        // A range that ends at -1.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x00000000,
+                end: 0xffffffff,
+            }))
+        );
+
+        // A range end.
+        assert_eq!(ranges.next(), Ok(None));
+
+        // An offset at the end of buf.
+        let mut ranges = rnglists
+            .ranges(RangeListsOffset(buf.len()), version, 4, 0x01000000)
+            .unwrap();
+        assert_eq!(ranges.next(), Ok(None));
+    }
+
+    #[test]
+    fn test_ranges_64() {
+        let start = Label::new();
+        let first = Label::new();
+        let section = Section::with_endian(Endian::Little)
+            // A range before the offset.
+            .mark(&start)
+            .L64(0x10000).L64(0x10100)
+            .mark(&first)
+            // A normal range.
+            .L64(0x10200).L64(0x10300)
+            // A base address selection followed by a normal range.
+            .L64(0xffffffffffffffff).L64(0x02000000)
+            .L64(0x10400).L64(0x10500)
+            // An empty range followed by a normal range.
+            .L64(0x10600).L64(0x10600)
+            .L64(0x10800).L64(0x10900)
+            // A range that starts at 0.
+            .L64(0).L64(1)
+            // A range that ends at -1.
+            .L64(0xffffffffffffffff).L64(0x00000000)
+            .L64(0).L64(0xffffffffffffffff)
+            // A range end.
+            .L64(0).L64(0)
+            // Some extra data.
+            .L64(0);
+
+        let buf = section.get_contents().unwrap();
+        let debug_ranges = DebugRanges::new(&buf, LittleEndian);
+        let debug_rnglists = DebugRngLists::new(&[], LittleEndian);
+        let rnglists = RangeLists::new(debug_ranges, debug_rnglists).unwrap();
+        let offset = RangeListsOffset((&first - &start) as usize);
+        let version = 4;
+        let mut ranges = rnglists.ranges(offset, version, 8, 0x01000000).unwrap();
+
+        // A normal range.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x01010200,
+                end: 0x01010300,
+            }))
+        );
+
+        // A base address selection followed by a normal range.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x02010400,
+                end: 0x02010500,
+            }))
+        );
+
+        // An empty range followed by a normal range.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x02010800,
+                end: 0x02010900,
+            }))
+        );
+
+        // A range that starts at 0.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x02000000,
+                end: 0x02000001,
+            }))
+        );
+
+        // A range that ends at -1.
+        assert_eq!(
+            ranges.next(),
+            Ok(Some(Range {
+                begin: 0x0,
+                end: 0xffffffffffffffff,
+            }))
+        );
+
+        // A range end.
+        assert_eq!(ranges.next(), Ok(None));
+
+        // An offset at the end of buf.
+        let mut ranges = rnglists
+            .ranges(RangeListsOffset(buf.len()), version, 8, 0x01000000)
+            .unwrap();
+        assert_eq!(ranges.next(), Ok(None));
+    }
+
+    #[test]
+    fn test_ranges_invalid() {
+        let section = Section::with_endian(Endian::Little)
+            // An invalid range.
+            .L32(0x20000).L32(0x10000)
+            // An invalid range after wrapping.
+            .L32(0x20000).L32(0xff010000);
+
+        let buf = section.get_contents().unwrap();
+        let debug_ranges = DebugRanges::new(&buf, LittleEndian);
+        let debug_rnglists = DebugRngLists::new(&[], LittleEndian);
+        let rnglists = RangeLists::new(debug_ranges, debug_rnglists).unwrap();
+        let version = 4;
+
+        // An invalid range.
+        let mut ranges = rnglists
+            .ranges(RangeListsOffset(0x0), version, 4, 0x01000000)
+            .unwrap();
+        assert_eq!(ranges.next(), Err(Error::InvalidAddressRange));
+
+        // An invalid range after wrapping.
+        let mut ranges = rnglists
+            .ranges(RangeListsOffset(0x8), version, 4, 0x01000000)
+            .unwrap();
+        assert_eq!(ranges.next(), Err(Error::InvalidAddressRange));
+
+        // An invalid offset.
+        match rnglists.ranges(RangeListsOffset(buf.len() + 1), version, 4, 0x01000000) {
+            Err(Error::UnexpectedEof) => {}
+            otherwise => panic!("Unexpected result: {:?}", otherwise),
+        }
     }
 }
